services/ingestion_service/requirements.txt
Flask
psycopg2-binary
confluent-kafka
minio
pypdf
nltk
requests